- set gyro scale to maximum (16g, 2000d/sec) - set heading format to degrees - modified web upload and live web application

+++ software/dotnet/GroundControl.Core/TelemetryData.cs
@@ -30,7 +30,7 @@ namespace GroundControl.Core
         /// <summary>
         /// The heading in radians.
         /// </summary>
-        public float Heading { get; set; }
+        public short Heading { get; set; }
 
         /// <summary>
         /// The speed over ground in m/s.
